% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.r
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(kobodata, filename = "fundata.csv", outdir = getwd())
}
\arguments{
\item{kobodata}{name of the funaction-kobo data file}

\item{filename}{name of the output file}

\item{outdir}{directory path where to store the file

OUTPUT}
}
\value{
a prepared data frame (also writes data on disk to outdir)
}
\description{
DESCRIPTION
the function reads funaction kobotoolbox csv file, removes \if{html}{\out{<not 
approved>}} data and photo data columns, simplifies variable names,
and returns the resulting dataframe object
}
\details{
PARAMETERS
}
