% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.r
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(
  kobodata,
  outfname = "fundata.csv",
  projectname = "funaction",
  rm_photo_data = TRUE,
  rm_data_check_columns = TRUE,
  take_only_approved_data = FALSE
)
}
\arguments{
\item{kobodata}{name of the funaction-kobo data file}

\item{outfname}{name of the output file}

\item{projectname}{filter by project name  (e.g., funaction)
Use "all" to select all projects}

\item{rm_photo_data}{remove data related to photos. Default TRUE}

\item{rm_data_check_columns}{remove boolean data check variables.
Default TRUE}

\item{take_only_approved_data}{not approved data is discarded
Default FALSE

OUTPUT}
}
\value{
a prepared data frame (also writes data on disk to outdir)
}
\description{
DESCRIPTION
the function reads funaction kobotoolbox csv file, removes \if{html}{\out{<not 
approved>}} data and photo data columns, simplifies variable names,
and returns the resulting dataframe object
}
\details{
PARAMETERS
}
